+++ docs/conf.py
@@ -57,7 +57,7 @@ author = "Matt Wright & Chris Wagner"
 # built documents.
 #
 # The short X.Y version.
-version = "4.1.5"
+version = "4.1.6"
 # The full version, including alpha/beta/rc tags.
 release = version