+++ tests/test_datastore.py
@@ -436,8 +436,9 @@ def test_uuid(app, request, tmpdir, realdburl):
         db.create_all()
 
     def tear_down():
-        db.drop_all()
-        _teardown_realdb(db_info)
+        with app.app_context():
+            db.drop_all()
+            _teardown_realdb(db_info)
 
     request.addfinalizer(tear_down)