+++ tests/test_registerable.py
@@ -116,9 +116,9 @@ def test_registerable_flag(clients, app, get_message):
 
 @pytest.mark.confirmable()
 @pytest.mark.babel()
+@pytest.mark.app_settings(babel_default_locale="fr_FR")
 def test_xlation(app, client, get_message_local):
     # Test form and email translation
-    app.config["BABEL_DEFAULT_LOCALE"] = "fr_FR"
     assert check_xlation(app, "fr_FR"), "You must run python setup.py compile_catalog"
 
     response = client.get("/register", follow_redirects=True)